+++ ui/src/kapacitor/components/config/SMTPConfig.tsx
@@ -1,4 +1,5 @@
-import React, {PureComponent} from 'react'
+import _ from 'lodash'
+import React, {PureComponent, ChangeEvent} from 'react'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 
 interface Properties {
@@ -18,6 +19,7 @@ interface Config {
     password: boolean
     from: string
     to: string | string[]
+    enabled: boolean
   }
 }
 
@@ -30,6 +32,7 @@ interface Props {
 
 interface State {
   testEnabled: boolean
+  enabled: boolean
 }
 
 @ErrorHandling
@@ -45,11 +48,17 @@ class SMTPConfig extends PureComponent<Props, State> {
     super(props)
     this.state = {
       testEnabled: this.props.enabled,
+      enabled: _.get(this.props, 'config.options.enabled', false),
     }
   }
 
   public render() {
-    const {host, port, from, username, password, to} = this.props.config.options
+    const {host, port, from, username, password, to} = _.get(
+      this.props,
+      'config.options',
+      {}
+    )
+    const {enabled} = this.state
 
     return (
       <form onSubmit={this.handleSubmit}>
@@ -127,6 +136,18 @@ class SMTPConfig extends PureComponent<Props, State> {
           />
         </div>
 
+        <div className="form-group col-xs-12">
+          <div className="form-control-static">
+            <input
+              type="checkbox"
+              id="disabled"
+              checked={enabled}
+              onChange={this.handleEnabledChange}
+            />
+            <label htmlFor="disabled">Configuration Enabled</label>
+          </div>
+        </div>
+
         <div className="form-group form-group-submit col-xs-12 text-center">
           <button
             className="btn btn-primary"
@@ -138,7 +159,7 @@ class SMTPConfig extends PureComponent<Props, State> {
           </button>
           <button
             className="btn btn-primary"
-            disabled={!this.state.testEnabled}
+            disabled={!this.state.testEnabled || !enabled}
             onClick={this.props.onTest}
           >
             <span className="icon pulse-c" />
@@ -147,6 +168,11 @@ class SMTPConfig extends PureComponent<Props, State> {
         </div>
       </form>
     )
+  }
+
+  private handleEnabledChange = (e: ChangeEvent<HTMLInputElement>) => {
+    this.setState({enabled: e.target.checked})
+    this.disableTest()
   }
 
   private handleSubmit = async e => {
@@ -159,6 +185,7 @@ class SMTPConfig extends PureComponent<Props, State> {
       to: this.to.value ? [this.to.value] : [],
       username: this.username.value,
       password: this.password.value,
+      enabled: this.state.enabled,
     }
     const success = await this.props.onSave(properties)
     if (success) {